Add Power Control and Use SPI0 Only
As SPI1 exists, it seems an easy way to program two chips that
don't have chip select or enable pins.  BUT, avrdude, which I'm
using to program the AVRs, only handles the baud rate correctly
on SPI0.  That being the case, enhance the circuit so that two
GPIO pins are used to power on or off the AVRs.

Signed-off-by: John Jacques <eruditelite@gmail.com>